Refactor for candlestick data
This commit includes a variety of updates to get FundRunner working with the new candlestick schema, outlined in the poloniex-ticker-history repo.

These updates required a few structural changes. For one thing, charts no longer refer to currencies, but to *markets* - candlestick data always refer to the rate of some currency (the fiat, BTC for now) and the currency of interest.

For another, coinstore now must know the trading duration of the strategy, so that it can query for the proper candlestick data. Since this couples the coinstore more tightly with particular strategies, it may eventually make more sense to have strategies manage coinstores internally. (This mayalso help if we end up running backtests in parallel).

Finally, I reran all three of our strategies against all the data in our corpus. PeakRider at a twenty-four-hour duration seems to offer the highest Sterling ratio over all time.

Minor changes:

- Thorough backtest runner is now called backtest.py

- Simple example now called example.py

- Redid python notebook with new backtest results

- Updated .gitignore