Update FundRunner for new historical data
Previously, backtesting used some OHLCV format we marshalled the scraped Poloniex data into.  Since the live trading algorithm is now using Poloniex's API to make decisions, it makes sense to use this API format the whole way through, without marshalling any data.

This PR updates the backtesting code to use scrapedData from the historical corpus.

All chart data is now 15 minute intervals. By default, we trade every 15 minutes.  This removes some of the complexity associated with having variable durations. For example, in PeakRider, the estimated moving averages should change size depending on the candlestick duration, so that they cover the same amount of actual time.  Since we overlooked this error, errors like this are probably easy to overlook in general, and can be removed completely by sticking to 15 minute intervals in both live-trading and backtesting scenarios.  If we ever want to add this functionality back, we can always group scraped chart readings together and perform some data marshalling on the strategy side. At least this way, the strategy has to be aware of the transformation, versus having it hidden.

This PR breaks the backtest.py / backtest-conditions.json way of doing backtests with different controls. We'll have to consider better ways of doing this in the future.

Also, backtesting is slow as shit. How can we speed it up? Anyone know how to profile Python code?