Replace Purchase with ProposedTrade
Purchases are weird, because the name implies they have been completed. In reality, Purchases are really just plans, or hopes, of what might happen on the market.

This commit replaces Purchases with a new type: ProposedTrade. ProposedTrades have a from_coin (we are trading from), and a to_coin (we are trading to). Objects of this type can calculate the price of the to_coin, given current charts. Finally, when callers are ready, they can `set_bid_amount()`, passing in how much from_coin they wish to sell. The object will automatically estimate how much we should receive in return (ask_amount), incorporating the estimated price and estimated fees.

- Strategies now implement a propose_trades() method as their main API

- Step() passes time to propose_trades()

- All tests passing.